Clean Up Publish Process (#21)
## TLDR;
Cleans up the publish process and removes the `dart_jsx` package from
the list of packages to be published.

## What Does This Do?
This change ensures that the `dart_jsx` package is no longer included in
the publishing process. This prevents it from being built or published.

## Brief Details?
The `dart_jsx` package has been removed from the `packageDeps` and
`publishOrder` lists within the `prepare_publish.dart` file. This
exclusion effectively removes the package from the automated publishing
workflow.

## How Do The Tests Prove The Change Works?
There are no specific tests included in this change, but the removal of
`dart_jsx` prevents it from being included in any tests related to the
publishing process.